A good evening/night here. Came back from the field quite late but have worked on a few classes - moving up the progress of the project. With a few challenges that we can work together on. Earlier I mentioned the issue with the mouse event/capture. it still doesn't work, I think if you fixed it didn't commit to the git. As soon as this mouse MouseListener issue is fixed our project is ready and all set. What I think however needs to be done to fix this which we can look at is:
    a. the mouse events should trigger a repaint() library
    b. the combo box should be switched, if-else﻿

What do you think?

In summary, this is what I did tonight:

a. Mouse Events.
I tried to deal with the design of mouse trigger events enabling the user trigger mouse events such as users to click images to draw elements on the screen when the users click on the panel the users simultaneously draw different shapes.

b. Graphical User Interface
I have added a few functionalities to the project such as erase, redo and undo.

c. Polymorphism.

In order to keep up with the instructions for the assignment, i have created and implemented some implementation classes for My Line, My Rectangle & implementation for all the shape & filling respective colors.

Challenges:

a. Mouse Pointer Instances

As mentioned above, I am having issues with mouse clicks, tried to work on creating images with mouse on-click.

b. An implementation framework for the buttons and erase on-click events.

Once these are done, I think it will be fine - I have run similar in NetBeans and it seems much easier for the first challenge, for some reason with IntelliJ bit challenging.

Best,