Awesome, and to top up the cake, I have added a few extra functionalities to the project for bonus points - such as” The size of the shape changes as one drags the mouse", have added undo which is using linked list to create a stack, I’ve also added the bit on how we did the polymorphism and inheritance and explain some of the methods in JavaPaintApp.java. Now you can clear the last shape or clear all. Clear will clear one by one while clear all will clear everything on the board.
I also saw that on the main instruction of the instructions, we should draw in a continuous motion, and also, I have made the dragging to set the size of the shape. Before you would drag but will only see the shape when you let go but now I have made it possible to show and reduce or increase the size of the shapes as you draw, you can also erase the shapes on the panel one by one, or clear all. To achieve the drag and resize the shapes I had to use the code to build the GUI, I created the panel, using component method in JavaPaintApp. In order to take advantage of the repaint function that is already in java, I have remove the panels as we did initially and used component method. Also, I have created three new classes for that.

Ive also refactored the naming from the cordinate1 etc to x1 etc.

Also note that I have modified the interface a bit to reflect the new functionality and visibility.

Lets have a final skype call and I will commit the changes,

Also attached I have modified the documentation and added the new functionalities and give emphasis on how we did the polymorphism, inheritance and our methods. After the call will modify the screenshots and we are good